% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{get_overview}
\alias{get_overview}
\title{Get overview of topics/runs}
\usage{
get_overview(stm_models)
}
\arguments{
\item{stm_models}{A list of stm model objects from
\code{stm::selectModels()}.}
}
\description{
Get overview of topics/runs
}
