% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_semcoh.R
\name{get_stats}
\alias{get_stats}
\title{Extract semantic coherence and exclusivity form a list of models}
\usage{
get_stats(stm_models)
}
\arguments{
\item{stm_models}{A list of stm model object created by
\code{selectModels()}.}
}
\description{
Extract semantic coherence and exclusivity form a list of models
}
