% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.R
\name{plot_conv_all}
\alias{plot_conv_all}
\title{Plot convergence for a list of stm models}
\usage{
plot_conv_all(stm_models)
}
\arguments{
\item{stm_models}{A list of stm model (from \code{selectModels()}).}
}
\description{
Plot convergence for a list of stm models
}
