% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frex.R
\name{plot_frex}
\alias{plot_frex}
\title{Alternative plotting of FREX}
\usage{
plot_frex(fit, n = 15)
}
\arguments{
\item{fit}{An ‘STM’ model object.}

\item{n}{Number of words to print.}
}
\description{
Alternative plotting of FREX
}
