% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gini.R
\name{plot_gini_docs}
\alias{plot_gini_docs}
\title{Plot gini coefficient of topics within documents}
\usage{
plot_gini_docs(fit)
}
\arguments{
\item{fit}{An \code{STM} model object.}
}
\description{
Plot gini coefficient of topics within documents
}
