% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gini.R
\name{plot_gini_theta}
\alias{plot_gini_theta}
\title{Plot gini of documents within topics}
\usage{
plot_gini_theta(fit)
}
\arguments{
\item{fit}{An \code{STM} model object.}
}
\description{
Plot gini of documents within topics
}
