% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plot_main_header}
\alias{plot_main_header}
\title{Plot the header on the first in plot_models.}
\usage{
plot_main_header(title = NULL)
}
\arguments{
\item{title}{Title to print over session info.}
}
\description{
Plot the header on the first in plot_models.
}
