% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot the overview used in plot_models}
\usage{
plot_overview(fit, overview, run_n_current)
}
\arguments{
\item{fit}{A stm model object.}

\item{overview}{The output from \code{get_overview()}, which in
turn where fed a list of stm model object from
\code{selectModels()}.}

\item{run_n_current}{Current run to highlight.}
}
\description{
Plot the overview used in plot_models
}
