% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prev.R
\name{plot_prev}
\alias{plot_prev}
\title{Alternative plotting of topic prevelance}
\usage{
plot_prev(fit)
}
\arguments{
\item{fit}{An ‘STM’ model object.}
}
\description{
Alternative plotting of topic prevelance
}
