% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_semcoh.R
\name{plot_semcoh}
\alias{plot_semcoh}
\title{Plot semantic coherence and exclusivity for one model.}
\usage{
plot_semcoh(stats, mod, n_run = NULL)
}
\arguments{
\item{stats}{Output from \code{get_stats()}, which in turn were fed
a list of stm model object created by \code{selectModels()}.}

\item{mod}{Which n model to print (corresponding to it's place in
the list of model for \code{get_stats()}.}

\item{n_run}{= Which run of the model to highlight.}
}
\description{
Plot semantic coherence and exclusivity for one model.
}
