% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_semcoh.R
\name{plot_semcoh_all}
\alias{plot_semcoh_all}
\title{Plot semantic coherence for all stm models}
\usage{
plot_semcoh_all(stats)
}
\arguments{
\item{stats}{Output from \code{get_stats()}, which in turn were fed
a list of stm models.}
}
\description{
Plot semantic coherence for all stm models
}
