% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thoughts.R
\name{plot_thoughts}
\alias{plot_thoughts}
\title{Alternative plotting of most representative texts}
\usage{
plot_thoughts(fit, texts, split_width = 62, n_sentences = 25,
  cutoff = 190)
}
\arguments{
\item{fit}{An ‘STM’ model object.}

\item{texts}{A character vector where each entry contains the text
of a document.  Must be in the same order as the documents
object. NOTE: This is not the documents which are passed to ‘stm’
and come out of ‘prepDocuments’, this is the actual text of the
document.}

\item{split_width}{The character width to split texts/thoughts into
a new line.}

\item{n_sentences}{The number of sentences to print per topic.}

\item{cutoff}{The number of characters before the text/thought
should be cut off with \code{(...)}.}
}
\description{
Alternative plotting of most representative texts
}
