% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plot_title}
\alias{plot_title}
\title{Plot the title used in plot_models}
\usage{
plot_title(fit, run_n_current)
}
\arguments{
\item{fit}{A stm model object.}

\item{run_n_current}{Current run.}
}
\description{
Plot the title used in plot_models
}
