% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_models.R
\name{print_models}
\alias{print_models}
\title{Print a list of models to a multi-page pdf file.}
\usage{
print_models(stm_models, texts, file = "printed_stm.pdf", title = NULL,
  verbose = TRUE)
}
\arguments{
\item{stm_models}{Either a list of stm model objects from
\code{selectModels()} or the output of \code{manyTopics()}.}

\item{texts}{A character vector where each entry contains the text
of a document. Must be in the same order as the documents
object. NOTE: This is not the documents which are passed to ‘stm’
and come out of ‘prepDocuments’, this is the actual text of the
document.}

\item{file}{either a character string naming a file or a connection
open for writing.}

\item{title}{Title to printed on top of the first page.}

\item{verbose}{If it should show a progress bar in the terminal.}
}
\description{
Print a list of models to a multi-page pdf file.
}
